// src/app/connectors/api/model/tokenResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TokenResponse { 
    token?: string;
    expiresAt?: number;
    isNewUser?: boolean;
}